requirements.txt
antropy==0.1.6
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.9.0
h5py==3.11.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupytext==1.16.4
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
ml-dtypes==0.4.0
mne==1.8.0
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.24.0
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primefac==2.0.12
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyEDFlib==0.1.38
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.7.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
seizure_data_processing @ git+https://github.com/sderooij/seizure_data_processing.git@0fcb5cc2c860415dde24d0f687b592cc4021d21f
six==1.16.0
stack-data==0.6.3
stochastic==0.6.0
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
